# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GSDB API",
    "version": "0.1.0",
    "description": "\n# GSDB API Documentation\n\n**Authentication:** To use the API, you must provide an API key as a Bearer token.  \nThis token can be obtained in your user profile edit page on the website: [User Area](https://gsdb-web.onrender.com/user-area).  \nInclude it in the header of every request:\n\n`Authorization: Bearer <API_KEY>`\n\nFor any other information, like how to use query parameters or the different database fields, you can visit the [GSDB Wiki](https://github.com/Game-Save-DataBase/GSDB_Server/wiki)\n"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_KEY"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/games": {
      "get": {
        "summary": "Search games",
        "description": "Search games by multiple query filters. Supports MongoDB query operands. Can search local database and external IGDB API.",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Search keyword (optional)"
          },
          {
            "in": "query",
            "name": "platformID",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by platform ID (optional)"
          },
          {
            "in": "query",
            "name": "release_date",
            "schema": {
              "type": "string"
            },
            "description": "Filter by release date (optional)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Maximum number of results (optional)"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            },
            "description": "Offset for pagination (optional)"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "object"
            },
            "description": "Sort field and order (optional)"
          },
          {
            "in": "query",
            "name": "external",
            "schema": {
              "type": "boolean"
            },
            "description": "If true, searches external IGDB API when local results are insufficient (default: true)\n"
          },
          {
            "in": "query",
            "name": "complete",
            "schema": {
              "type": "boolean"
            },
            "description": "Whether to include full game info (default: true)\n"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of game objects or single game"
          },
          "204": {
            "description": "No coincidences found"
          },
          "400": {
            "description": "Invalid query"
          }
        }
      },
      "delete": {
        "summary": "Delete a game",
        "description": "Delete a game by its ID. Requires authentication and dev mode.",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Game ID to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Game deleted successfully"
          },
          "400": {
            "description": "Missing or invalid game ID"
          },
          "404": {
            "description": "Game not found"
          }
        }
      }
    },
    "/games/search": {
      "get": {
        "summary": "Search games by keyword",
        "description": "Returns games matching a keyword search. Supports limit, offset, platform filter, release date, and sorting.",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Search keyword"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Limit number of results"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            },
            "description": "Offset for pagination"
          },
          {
            "in": "query",
            "name": "platformID",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by platform ID"
          },
          {
            "in": "query",
            "name": "release_date",
            "schema": {
              "type": "string"
            },
            "description": "Filter by release date"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "object"
            },
            "description": "Sort field and order"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of games matching search"
          },
          "204": {
            "description": "No matches found"
          },
          "400": {
            "description": "Invalid query"
          }
        }
      }
    },
    "/games/igdb": {
      "post": {
        "summary": "Add games from IGDB",
        "description": "Adds one or multiple games to the database using IGDB IDs or a range of IGDB IDs. Dev mode only.",
        "tags": [
          "Games"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "IGDB_ID": {
                    "type": "integer",
                    "description": "Single IGDB game ID"
                  },
                  "IGDB_ID_INIT": {
                    "type": "integer",
                    "description": "Start of IGDB ID range"
                  },
                  "IGDB_ID_END": {
                    "type": "integer",
                    "description": "End of IGDB ID range"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Game(s) added successfully"
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "No games found"
          }
        }
      }
    },
    "/games/dev/wipe": {
      "delete": {
        "summary": "Wipe all games",
        "description": "Deletes all games in the database. Development mode only.",
        "tags": [
          "Games"
        ],
        "responses": {
          "200": {
            "description": "All games deleted with deleted count"
          },
          "500": {
            "description": "Server error while wiping games"
          }
        }
      }
    },
...
  },
...
}